switch/requirements.txt
Flask==0.12
paho-mqtt==1.2